+++ src/main/java/com/N26/robotfactory/domain/model/ResponseRobotFactory.java
@@ -2,6 +2,8 @@ package com.N26.robotfactory.domain.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -9,5 +11,5 @@ import lombok.*;
 @Builder
 public class ResponseRobotFactory {
     private String order_id;
-    private Double total;
+    private BigDecimal total;
 }
